refactoring all classes, files, packages
