+++ app/src/main/java/com/example/answer/github/data/GithubData.kt
@@ -1,4 +1,4 @@
-package com.example.answer.github.room
+package com.example.answer.github.data
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
