+++ app/src/main/java/com/example/answer/github/data/GithubRepo.kt
@@ -1,4 +1,4 @@
-package com.example.answer.github.room
+package com.example.answer.github.data
 
 import com.google.gson.annotations.SerializedName
