+++ app/src/main/java/com/example/answer/github/data/source/local/GithubDatabase.kt
@@ -1,9 +1,11 @@
-package com.example.answer.github.room
+package com.example.answer.github.data.source.local
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.answer.github.data.FavoriteData
+import com.example.answer.github.data.GithubData
 
 @Database(entities = [GithubData::class, FavoriteData::class], version =2)
 abstract class GithubDatabase: RoomDatabase() {
