+++ app/src/main/java/com/example/answer/github/view/GithubSearchAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.answer.github.recyclerview
+package com.example.answer.github.view
 
 import android.view.LayoutInflater
 import android.view.View
@@ -9,15 +9,14 @@ import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.answer.R
-import com.example.answer.github.GithubActivity
-import com.example.answer.github.GithubViewModel
-import com.example.answer.github.room.GithubData
+import com.example.answer.github.viewmodel.GithubViewModel
+import com.example.answer.github.data.GithubData
 
 
-class GithubLikeAdapter : RecyclerView.Adapter<GithubLikeAdapter.ViewHolder>() {
+class GithubSearchAdapter : RecyclerView.Adapter<GithubSearchAdapter.ViewHolder>() {
     private var contacts: List<GithubData> = listOf()
     private lateinit var viewModel: GithubViewModel
-    private lateinit var view: GithubActivity
+    private lateinit var view : GithubActivity
 
     override fun onCreateViewHolder(parent: ViewGroup, i: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.github_item, parent, false)
@@ -42,7 +41,6 @@ class GithubLikeAdapter : RecyclerView.Adapter<GithubLikeAdapter.ViewHolder>() {
         fun bind(githubData: GithubData) {
             nameTv.text = githubData.name
             scoreTv.text = githubData.score.toString()
-            Glide.with(view.applicationContext).load(githubData.avatar_url).into(profileImage)
 
             if (githubData.favorite != 0) {
                 checkFavorite()
@@ -59,6 +57,8 @@ class GithubLikeAdapter : RecyclerView.Adapter<GithubLikeAdapter.ViewHolder>() {
                     viewModel.updateList(0, githubData.name)
                 }
             }
+
+            Glide.with(view.applicationContext).load(githubData.avatar_url).into(profileImage)
         }
 
         private fun checkFavorite() {
