+++ app/src/main/java/com/example/answer/github/view/GithubViewPagerAdapter.kt
@@ -1,12 +1,10 @@
-package com.example.answer.github
+package com.example.answer.github.view
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
 import com.example.answer.github.fragments.LikeFragment
 import com.example.answer.github.fragments.SearchFragment
-import com.example.answer.github.recyclerview.GithubLikeAdapter
-import com.example.answer.github.recyclerview.GithubSearchAdapter
 
 class GithubViewPagerAdapter(
     fm: FragmentManager?
