+++ app/src/main/res/layout/activity_github.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".github.GithubActivity">
+    tools:context=".github.view.GithubActivity">
 
     <data>
         <variable
